requirements.txt
alembic==0.8.5
amqp==1.4.9
anyjson==0.3.3
autoenv==1.0.0
beautifulsoup4==4.4.1
billiard==3.3.0.23
blinker==1.4
boto==2.42.0
celery==3.1.24
cffi==1.8.2
click==6.6
contextlib2==0.5.4
coverage==3.7.1
cryptography==1.5
cycler==0.9.0
dnet==1.12
ecdsa==0.10
enum34==1.1.6
Fabric==1.8.2
Facebook==0.0
facebook-sdk==2.0.0
Flask==0.10.1
Flask-Admin==1.4.2
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
funcsigs==0.4
GDAL==1.11.5
gunicorn==19.3.0
hashpumpy==1.2
idna==2.1
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.37
Mako==1.0.4
MarkupSafe==0.23
matplotlib==1.5.1
mediameter-cliff==1.2.0
mock==1.3.0
MySQL-python==1.2.5
nltk==3.2
nose==1.3.7
numpy==1.11.2
paramiko==1.12.2
pbr==1.10.0
Pillow==3.4.1
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pyparsing==2.0.6
python-dateutil==2.4.2
python-editor==0.5
pytz==2015.7
raven==5.27.1
redis==2.10.5
requests==2.9.1
rq==0.5.6
simplejson==3.8.2
six==1.10.0
SQLAlchemy==1.0.8
stevedore==1.14.0
vboxapi==1.0
virtualenv==15.0.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
Werkzeug==0.10.4
WTForms==2.0.2
